requirements.txt
Flask==3.0.0
google-api-python-client==2.112.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
gunicorn==21.2.0
Pillow==10.1.0
Flask-Limiter==3.5.0
bleach==6.1.0